+++ design-system/datetime-picker/src/components/__tests__/vr-tests/time-picker-label-options.vr.tsx
@@ -1,0 +1,7 @@
+import { snapshot } from '@af/visual-regression';
+
+import LabelExamples from '../../../../examples/130-date-time-label-examples';
+
+snapshot(LabelExamples, {
+  drawsOutsideBounds: true,
+});