+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-mobile-bridge
 
+## 193.0.5
+
+### Patch Changes
+
+- Updated dependencies
+
+## 193.0.4
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 193.0.3
 
 ### Patch Changes